+++ cvd_src/fast/fast_11_detect.cxx
@@ -2581,7 +2581,7 @@ void fast_corner_detect_plain_11(const BasicImage<byte>& i, vector<ImageRef>& co
 			  continue;
 
 			success:
-				corners.push_back(ImageRef(cache_0-line_min, y));
+				corners.push_back(ImageRef(static_cast<int>(cache_0-line_min), y));
 		}
 	}
 }
